Refactored Budgets Module
- Removed old Views
- Removed tests for old views
- Removed old URLs
- Removed old Templates
- Refactored view.py into a directory with the budget views and the Money Distribution Views
- Moved new urls from rest_api to budget
