+++ application/budget/views/__init__.py
@@ -1,0 +1,2 @@
+from budget_views import *
+from money_distribution_views import *
