+++ application/static_border_stations/templates/static_border_stations/borderstation_form.html
@@ -1,8 +1,5 @@
 <div class="row well well-sm">
   <h1>Details
-    {% if request.user.permission_budget_manage and not is_create %}
-      <a class="btn btn-success pull-right" href="{% url 'budget_create_api' form.instance.id %}">New Budget Calculation Form</a>
-    {% endif %}
 
   {% if request.user.permission_border_stations_delete%}
       {% verbatim %}
